S003 fix : 회의 중 수정사항. 기타 이름 및 Error 수정
index key
usecase name

+++ Gustav/Domain/Entity/Workspace.swift
@@ -11,8 +11,9 @@ import Foundation
 struct Workspace {
     let id: UUID            // 워크스페이스 ID
     let userId: UUID        // 사용자 ID
-    let indexKey: Decimal    // 정렬 기준 값
+    let indexKey: Int   // 정렬 기준 값
     let name: String        // 워크스페이스 이름
     let createdAt: Date     // 생성일
     let updatedAt: Date     // 수정일
 }
+
